requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=2.0.0
supabase>=1.0.3
python-dotenv>=0.19.0 